+++ cpp/src/io/parquet/writer_impl.cu
@@ -122,6 +122,7 @@ class parquet_column_view {
   explicit parquet_column_view(size_t id,
                                column_view const &col,
                                const table_metadata *metadata,
+                               bool int96_timestamps,
                                cudaStream_t stream)
     : _col(col),
       _leaf_col(get_leaf_col(col)),
@@ -225,23 +226,23 @@ class parquet_column_view {
         _stats_dtype    = statistics_dtype::dtype_int32;
         break;
       case cudf::type_id::TIMESTAMP_SECONDS:
-        _physical_type  = Type::INT64;
+        _physical_type  = int96_timestamps ? Type::INT96 : Type::INT64;
         _converted_type = ConvertedType::TIMESTAMP_MILLIS;
         _stats_dtype    = statistics_dtype::dtype_timestamp64;
         _ts_scale       = 1000;
         break;
       case cudf::type_id::TIMESTAMP_MILLISECONDS:
-        _physical_type  = Type::INT64;
+        _physical_type  = int96_timestamps ? Type::INT96 : Type::INT64;
         _converted_type = ConvertedType::TIMESTAMP_MILLIS;
         _stats_dtype    = statistics_dtype::dtype_timestamp64;
         break;
       case cudf::type_id::TIMESTAMP_MICROSECONDS:
-        _physical_type  = Type::INT64;
+        _physical_type  = int96_timestamps ? Type::INT96 : Type::INT64;
         _converted_type = ConvertedType::TIMESTAMP_MICROS;
         _stats_dtype    = statistics_dtype::dtype_timestamp64;
         break;
       case cudf::type_id::TIMESTAMP_NANOSECONDS:
-        _physical_type  = Type::INT64;
+        _physical_type  = int96_timestamps ? Type::INT96 : Type::INT64;
         _converted_type = ConvertedType::TIMESTAMP_MICROS;
         _stats_dtype    = statistics_dtype::dtype_timestamp64;
         _ts_scale       = -1000;  // negative value indicates division by absolute value
@@ -586,6 +587,7 @@ writer::impl::impl(std::unique_ptr<data_sink> sink,
   : _mr(mr),
     compression_(to_parquet_compression(options.get_compression())),
     stats_granularity_(options.get_stats_level()),
+    int96_timestamps(options.is_enabled_int96_timestamps()),
     out_sink_(std::move(sink))
 {
 }
@@ -595,9 +597,10 @@ std::unique_ptr<std::vector<uint8_t>> writer::impl::write(
   const table_metadata *metadata,
   bool return_filemetadata,
   const std::string &column_chunks_file_path,
+  bool int96_timestamps,
   cudaStream_t stream)
 {
-  pq_chunked_state state{metadata, SingleWriteMode::YES, stream};
+  pq_chunked_state state{metadata, SingleWriteMode::YES, int96_timestamps, stream};
 
   write_chunked_begin(state);
   write_chunk(table, state);
@@ -629,7 +632,8 @@ void writer::impl::write_chunk(table_view const &table, pq_chunked_state &state)
     const auto current_id = parquet_columns.size();
 
     num_rows = std::max<uint32_t>(num_rows, col.size());
-    parquet_columns.emplace_back(current_id, col, state.user_metadata, state.stream);
+    parquet_columns.emplace_back(
+      current_id, col, state.user_metadata, state.int96_timestamps, state.stream);
   }
 
   if (state.user_metadata_with_nullability.column_nullable.size() > 0) {
@@ -700,9 +704,11 @@ void writer::impl::write_chunk(table_view const &table, pq_chunked_state &state)
         list_schema[nesting_depth * 2].name = "element";
         list_schema[nesting_depth * 2].repetition_type =
           col.leaf_col().nullable() ? OPTIONAL : REQUIRED;
-        list_schema[nesting_depth * 2].type           = col.physical_type();
-        list_schema[nesting_depth * 2].converted_type = col.converted_type();
-        list_schema[nesting_depth * 2].num_children   = 0;
+        auto const &physical_type           = col.physical_type();
+        list_schema[nesting_depth * 2].type = physical_type;
+        list_schema[nesting_depth * 2].converted_type =
+          physical_type == parquet::Type::INT96 ? ConvertedType::UNKNOWN : col.converted_type();
+        list_schema[nesting_depth * 2].num_children = 0;
 
         std::vector<std::string> path_in_schema;
         std::transform(
@@ -714,8 +720,10 @@ void writer::impl::write_chunk(table_view const &table, pq_chunked_state &state)
       } else {
         SchemaElement col_schema{};
         // Column metadata
-        col_schema.type           = col.physical_type();
-        col_schema.converted_type = col.converted_type();
+        auto const &physical_type = col.physical_type();
+        col_schema.type           = physical_type;
+        col_schema.converted_type =
+          physical_type == parquet::Type::INT96 ? ConvertedType::UNKNOWN : col.converted_type();
         // because the repetition type is global (in the sense of, not per-rowgroup or per
         // write_chunk() call) we cannot know up front if the user is going to end up passing tables
         // with nulls/no nulls in the multiple write_chunk() case.  so we'll do some special
@@ -1165,9 +1173,11 @@ std::unique_ptr<std::vector<uint8_t>> writer::write(table_view const &table,
                                                     const table_metadata *metadata,
                                                     bool return_filemetadata,
                                                     const std::string column_chunks_file_path,
+                                                    bool int96_timestamps,
                                                     cudaStream_t stream)
 {
-  return _impl->write(table, metadata, return_filemetadata, column_chunks_file_path, stream);
+  return _impl->write(
+    table, metadata, return_filemetadata, column_chunks_file_path, int96_timestamps, stream);
 }
 
 // Forward to implementation